Add support for obtaining inflection via lemminflect
Add an /inflections endpoint that takes a word (text) and part of speech tag
and returns the form of that word associated with the tag.

This commit defines a Word class that isn't any different from Document, but
its clearer to the user what a word is than a document expecting one word.

Signed-off-by: Moses Mendoza <mendoza.moses@gmail.com>